Initial effort to rewrite the integration specs to make them pass.
These haven't really passed for over a year and the project is hurting
for it.

This change makes the Lumberjack::Server a bit easier to test by
allowing easier access to accept a single connection. The monolithic
`run`  method wasn't a great idea anyway.

The current spec suite will write 3 lines to a file and will verify that
logstash-forwarder will read a file and forward its data to a lumberjack
server.

Fixes elastic/logstash-forwarder#366