Lumberjack apply back pressure to the producer
This pull request aim to applies the pipeline back pressure up to
logstash-forwarder (LSF) clients. In the current implementation the lumberjack
input create a new thread per LSF connection and start accepting new data right
away, if the pipeline is slow or there is a transient or a permanent error on
the output, each connection threads will block and remain in this state until
the pipeline unblock.

On the LSF side theses block will be detected as a connection timeout and
they will reconnect to logstash and restransmit the event frame,
Logstash will accept this new connection, read the data and block
again. LSF will never stop restransmitting data and logstash will create
threads until the process is out of memory.

This pull request bring a few features to the lumberjack input:

  * A Small buffered queue with a blocking timeout, a connection thread will
     never stay open forever.
  * A Circuit breaker, if there is too many timeout errors the breaker
    will open and logstash will start refusing new connection for a period
    of time.
  * Threads are now added to a cached threadpool backed by concurrent-ruby
    and java ThreadPoolExecutor (currently unbounded)

Fixes #12

Fixes #1