Change how we calculate the window size.
Window size was incorrectly caculated on the server side.
Causing the sequence number to be out of sync with the LSF, we were
using the first sequence number of a transmission to actually ack
a whole transmission.

Also removed the `window_size` option in the ruby client, because it
doesn't reflect the actual behavior of the ruby client. The client
doesn't do any buffering before transmitting the payload, so the client
should ack every send to not lose events.

Fixes #9