Fix a visibility issue accros threads
When the `#run` save a reference to the actual queue before using the
synchronous queue that reference is not garantee to be visible accross
all the threads. So if you call `#stop` its possible that the value is
nil. So this PR wrap the value inside a `Concurrent::MutexAtomicReference`
and use `#get` on it before inserting new items into the queue.