Make sure we stop the thread
The server was still reading the socket when the test was completed
making other tests fails.

Since we are mocking the connection to test the behavior of thread
connection handler calling `#stop` on the plugin wont correctly shutdown
the plugin. Until we have implemented a non blocking stack for beats
inputs I suggest we force to stop the server at the end of running the
tests.

Note in this change doesnt change the test for making sure the plugin
can be stopped.

Fixes #48

Fixes #49