TLS Client verification
This PR introduce client verification when using `ssl`, when using this
option you need to specify the **Certificate Authority** using the
`ssl_certificate_authorities` option, this configuration take a
CA to validate the client with it.

You also need to specify a `ssl_verify_mode`, If you want strict
ssl verification use `force_peer`, with this value if the client doesn't
provide a valid certificate the plugins will refuse the client.

On Filebeat you need to configure the client to declare a certificate using
theses options.

```
tls:
  certificate_authorities: ["/etc/ca.crt"]
  certificate: "/etc/client.crt"
  certificate_key: "/etc/client.key"
```

```
input {
  beats {
    ssl => true
    ssl_certificate_authorities => ["/etc/ca.crt"]
    ssl_certificate => "/etc/server.crt"
    ssl_key => "/etc/server.key"
    ssl_verify_mode => "force_peer
  }
}
```

This change also introduce testing with travis-ci and an integration
suite with filebeat for the SSL newer options.

Fix: #8

Fixes #56