Do not use the identity map unless we specify the multiline codec.
In some case when multiples FB clients are connected to the server and
and watching a lot of files we can hit the limit or the 20K identity
stream. This PR doesnt solve the 20K problems, but changes when the
identity map is used, we will only use it if the multiline codec is
defined in the configuration file. When its not the case the codec will
be reuse accross connection thread. We can do that since our codec are
stateless.

Fixes #70