Fix non deterministic tests with ports
Instead of just using a random number from a range, The
`Flores::Random.port` will try to setup a server on the specific port,
if everything is fine it will close the server and give you the port
number to use and we will know for sure that the port can be used.
The helper will try multiple port before giving up completely. (15
times)

This PR also make sure that all the tests are using random ports.

fixes #60

Fixes #71