Beats inputs Java rewrite
Motivation:

The core of Logstash is using more and more java instead of ruby for performance reason and scalability.
This PR introduce a lot of internal changes in the plugins and move from ruby to a pure java implementation using the Netty 4
Framework.

Also the connection model of this plugin has changed from a blocking to a non blocking implementation.
The changes should be backward compatible with the previous settings and will work for both LSF and Filebeat client.

Fixes #93