Gracefully recover from sending garbage data
This change make sure the state machien correctly recover from sending
garbage data over the write. The current strategy is to check to make
sure the frametype are correctly send if we received an unknown frame
type we will assume that the frame is poisonous, we will raise a custom
exception `InvalidFrameProtocol` with the associated bytes and will
terminate the connection since we cannot do anything with it.

Fixes: #100