Better handling of the payload size for v1 and v2 frame
It will raise an exception if the json payload size or the field count
is equal or inferior to 0. This commit also add a unit test for when the
v1 frames use UTF-8 characters for keys or the values to make sure they
are correctly encoded in the frame.

The assertions on the expected data was reinforced and will now accept
any name for the keys.

Fixes: #99

Fixes #106