IdleStateHandler is not `@Shareable` between channel.
This PR fix an issue introduced by #105 that was rejection new
connection on the server because the `IdleStateHandler` instance cannot
be shared between channels.

This also introduce a test to test concurrent connection on the server,
we never had a similar test before.

Gradle test will be also run explicitely on travis.

Fixes #111