Simplier SSL handling
This PR change how SSL is handled in this input, before we were using a
SSLKeyConverter to convert private key from PKCS7 to PKCS8 format, this conversion was
mostly only used for the integration specs because JRuby doesn't provide
an easy way to convert PKCS7 to PKCS8.

The code now delegate all the ssl handling directly to Netty's
SSLContext, integration suite still generate PKCS7 private key but we
use openssl binaries to convert the private key.