Fix an issue when specifying a passphrase that was raising a TypeError
The passphrase uses the `Password` class to encapsulate the sensitive
information and make sure we dont write it to the log by accident, to
get the original value the plugin must call `#value` on the object.

The Java layer was expecting a string but was receive a Password
instance and was throwing an exception.

This commit make sure we use password and add an integration test for
it.