Dont close the connection when the `IdleState.READER_IDLE` occur when processing a batch.
When logstash takes more time to process a batch than the `client_inactivity_timeout`,
the server was closing the connection in the middle of working on a batch. This commit changes the behavior,
if logstash is processing a batch and the event occur we wont drop the connection otherwise we will close it.

Fixes: #141