Fix an issue with when client authentification is on and CA contains multiples CA
Fix an issue when using client authentification and multiple CA defined
in the same **certificate_authorities** files. When the server read the
file it would only pick the first certificate in the file and ignore the
rest.

This problem could result in `unable to find valid certification path to
requested target` in Logstash and would refuse to let bet connect to it.

Fixes: #153