Correctly send the `client_inactivity_timeout` value to the beats server
The `client_inactivity_timeout` wasn't not correctly send to the Server
class so the server was using a default value instead. This PR add an
assertions on the value that the server class should receive, we have a
unit test in the java code that test the behavior of the timeout.

Fixes: #163

Fixes #164