Uses SO_LINGER on the socket option
When a idle client is detected the server will close the connection, if
you look at the TCP State the following scenario will happen.

- Server is set to FIN_WAIT_2
- Client is set to CLOSE_WAIT

Since FB never read from the socket it wont complete the handshake to
close the connection, change the socket option for `SO_LINGER` will
force the connection to be terminated from the server side.

On the FB side, the client will still think its connected until he need
to write something to the wire, when it happen it will just reconnect
to the server.

Using SO_LINGER is necessary in that case, because we have no way to
communicate to FB client that his connection was idle for too long, this
kind of two way messaging is not implemented yet in the protocol level.

Fixes #169