Breaking Change: No longer support multiline codec with beats input
When beats (in particular Filebeat) is not configured for multiline, incorrect groupings of data can result. When lines that should be grouped together are load balanced across multiple instances of Logstash, there is no means to properly reconstruct the grouping. It is also possible that lines that should be grouped together are not grouped properly on a single instance of Logstash due to the nature of beats batching. It is now required to handle multiline concerns in beats instead of Logstash. See https://www.elastic.co/guide/en/beats/filebeat/current/multiline-examples.html

Closes #201

Fixes #206