Rework the NioEventLoop handling in the beats input
We were using thow different NioEventLoop in the beats input, this was
causing problems in the context of back pressure. When the NioEventLoop
responsable of inserted to the queue was blocked we were still accepting
new connection instead of refusing them. Creating a large amount of
socket in the `CLOSE_WAIT` state.

This PR use a unique NioEventLoop, so when the thread pool is
blocked (by default is the 4 * number of core) we will queue new
connection respecting the `somaxconn` settings. (128 under linux, 200
under windows), LS will stop accepting new connection until the loop is
unblock and can accept new connection again.

On the other side this would result into `read io timeout error`