When an exception occur in the transmission close then flush buffered events.
when an exception was happening we were flushing the events to the
queue, this is was only problematic if the user was using a codec that
can buffer like the multiline codec, If the queue was under back
pressure, this would mean the thread would be lock on the insert to
the queue and the socket would never be correctly close.

Instead we now close the connection and propagate the exception, this
may also cause the thread to block, but the connection will be
gracefully stopped.