Improve Back pressure Handling
Previously, when Logstash queues were blocked, this plugin would
still accept new batches of events, but not send keep-alives back to
the client if the work queue was full. This would cause each
connected beats client to resend batches of work every time the TTL
was breached.

Also:
  Reduces the default number of beats executor threads from
4*number_of_processors to number_of_processors
  Includes fixes for #259

Fixes #299

Fixes #301