Add flag to enable disabling of 'host' field auto-population
Add `add_hostname` flag - when enabled, the `host` field will be populated
by the value supplied by the beat in the `hostname` field if that field is present.

The flag defaults to true to retain backwards compatibility with earlier versions of
the beats input.

The field is intentionally added as a deprecated field - auto populating the `host` field in
this way will be removed in future versions of the beats input.

Fixes #340