Fix sending of acks when sequence number of batch does not start with 1
Acknowledgements back to the sender were only being sent when the sequence
number of the message being handled was equal to the size of the batch.

This was causing a problem with the lumberjack output, as the underlying
jls-lumberjack library does not reset the sequence number to 1 for each
batch. This was stopping the beats input from correctly being able to ingest
files created by the lumberjack output.

This commit also fixes the BeatsHandler ack unit test, which previously did not
wait for, or test, the ack.

Fixes #342