Remove `java_import`s to avoid breaking tcp input
Since the release of `6.0.8`, when pipelines (or across multiple pipelines) with
tcp input and beats input, both using TLS, the TCP
input uses org.logstash.netty.SslContextBuilder defined here
instead of the desired SslContextBuilder from the Netty library, which
causes TCP pipelines to fail with `NoMethodError`.

This commit removes the (unnecessary) `java_imports` from the `beats.rb`
to avoid this.

This commit also pins devutils to `~>1.3` to fix the build.