Update/remove dependencies (#393)
* Update/remove dependencies

This commit updates the following dependencies:

Jackson-databind to `2.9.10.4`
Netty to `4.1.49.final`

Additionally, this commit removes the dependency on `tcnative` +
`boringssl` from #384, which enables the update of netty dependencies
while still allowing logstash->logstash communication via the lumberjack
output plugin. (`boringssl` removed compatibility with `CBC` ciphers, which
are still a prerequisite for lumberjack output plugin support at this time,

This commit also handles JCE restrictions

While it is been the default for Java since JDK 8u161 released in early 2018,
old versions of Java may not have the JCE unlimited strength jurisdiction policy
installed. This commit handles this case, warning the user that the policy is not
installed, and presenting a reduced set of default ciphers for use.