Separated the logic to drop incoming connections into specific handler
On new channel registration (that correspond to a new client connection),
verifies the direct memory stastus to understand if almost the totality max direct memory
is reached and also if the majoproity of that space is used by pinned byte buffers.
If the codition is verified that means direct memory avvailable is terminating, so no new
connection would help in the situation, and the incoming new connections are closed.