Logs a warn message on max order overshoot (#493)
Added logging the best Netty's maxOrder setting when big payload trigger an unpooled allocation.

Print a log line with the preferred Netty's maxOrder to accomodate the buffer size that exceeds the actual chunk size that would trigger an unpooled allocation.
It prints the log just one time across all plugin instances for each maxOrder required. So if plugin instance A reaches the limit of 4Mb (aka maxOrder 9) with a desired payload size of 7Mb it will report line about to use maxOrder 10. After this log line is printed, any other plugin instance wouldn't report it again, unless it needs an even bigger maxOrder not yet reported.
This PR also introduces a test spy of the Logger class to test the behaviour.

---------

Co-authored-by: João Duarte <jsvd@users.noreply.github.com>